edge-server/requirements.txt
annotated-types==0.6.0
paho-mqtt==2.1.0
pydantic==2.7.1
pydantic_core==2.18.2
typing_extensions==4.11.0
